requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyascii==0.3.2
    # via wagtail
arrow==1.3.0
    # via cookiecutter
asgiref==3.7.2
    # via
    #   django
    #   django-browser-reload
beautifulsoup4==4.11.2
    # via wagtail
binaryornot==0.4.4
    # via cookiecutter
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   cookiecutter
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   tqdm
cookiecutter==2.5.0
    # via python-webpack-boilerplate
defusedxml==0.7.1
    # via willow
distlib==0.3.8
    # via virtualenv
django==5.0
    # via
    #   -r requirements.in
    #   django-browser-reload
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-flags
    #   django-jquery-datatables
    #   django-modelcluster
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
    #   wagtailmedia
django-browser-reload==1.12.1
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.5
    # via wagtail
django-flags==5.0.13
    # via
    #   -r requirements.in
    #   wagtail-flags
django-jquery-datatables==1.0.8
    # via -r requirements.in
django-modelcluster==6.1
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-rest-framework==0.1.0
    # via django-jquery-datatables
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
djangorestframework==3.14.0
    # via
    #   django-rest-framework
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
filelock==3.13.1
    # via virtualenv
filetype==1.2.0
    # via willow
html5lib==1.1
    # via wagtail
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   cookiecutter
l18n==2021.3
    # via wagtail
lxml==4.9.4
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mysql-connector-python==8.2.0
    # via -r requirements.in
mysqlclient==2.2.1
    # via -r requirements.in
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.2
    # via -r requirements.in
openpyxl==3.1.2
    # via wagtail
packaging==23.2
    # via
    #   build
    #   pytest
pillow==10.1.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.14.0
    # via willow
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.in
protobuf==4.21.12
    # via mysql-connector-python
pygments==2.17.2
    # via rich
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
python-dotenv==1.0.0
    # via -r requirements.in
python-slugify==8.0.1
    # via cookiecutter
python-webpack-boilerplate==1.0.2
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via
    #   cookiecutter
    #   pre-commit
requests==2.31.0
    # via
    #   cookiecutter
    #   django-jquery-datatables
    #   wagtail
rich==13.7.0
    # via cookiecutter
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
telepath==0.3.1
    # via wagtail
text-unidecode==1.3
    # via python-slugify
tqdm==4.66.1
    # via -r requirements.in
types-python-dateutil==2.8.19.14
    # via arrow
tzdata==2023.3
    # via django
urllib3==2.1.0
    # via
    #   django-jquery-datatables
    #   requests
virtualenv==20.25.0
    # via pre-commit
wagtail==5.2.2
    # via
    #   -r requirements.in
    #   wagtail-color-panel
    #   wagtail-flags
    #   wagtailmedia
wagtail-color-panel==1.4.1
    # via -r requirements.in
wagtail-flags==5.3.0
    # via -r requirements.in
wagtailmedia==0.14.5
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
wheel==0.42.0
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in
willow[heif]==1.6.3
    # via
    #   wagtail
    #   willow

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
